refactor: Move shared test to base class
