+++ src/test/java/no/sikt/nva/pubchannels/handler/fetch/serialpublication/FetchSerialPublicationByIdentifierAndYearHandlerTest.java
@@ -1,6 +1,8 @@
 package no.sikt.nva.pubchannels.handler.fetch.serialpublication;
 
 import static java.net.HttpURLConnection.HTTP_OK;
+import static no.sikt.nva.pubchannels.TestConstants.API_DOMAIN;
+import static no.sikt.nva.pubchannels.TestConstants.CUSTOM_DOMAIN_BASE_PATH;
 import static no.sikt.nva.pubchannels.TestConstants.JOURNAL_TYPE;
 import static no.sikt.nva.pubchannels.TestConstants.SERIAL_PUBLICATION_PATH;
 import static no.sikt.nva.pubchannels.TestConstants.SERIES_TYPE;
@@ -11,7 +13,6 @@ import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 import com.google.common.net.MediaType;
 import java.io.IOException;
-import java.net.URI;
 import java.util.UUID;
 import no.sikt.nva.pubchannels.channelregistry.ChannelRegistryClient;
 import no.sikt.nva.pubchannels.channelregistrycache.db.service.CacheService;
@@ -22,6 +23,7 @@ import no.sikt.nva.pubchannels.handler.model.SerialPublicationDto;
 import no.sikt.nva.pubchannels.utils.AppConfig;
 import nva.commons.apigateway.GatewayResponse;
 import nva.commons.core.Environment;
+import nva.commons.core.paths.UriWrapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -55,8 +57,10 @@ public class FetchSerialPublicationByIdentifierAndYearHandlerTest extends
                                                                                          false));
         this.type = JOURNAL_TYPE;
         this.customChannelPath = SERIAL_PUBLICATION_PATH;
-        this.selfBaseUri = URI.create(
-            "https://localhost/publication-channels/" + SERIAL_PUBLICATION_PATH);
+        this.selfBaseUri = UriWrapper.fromHost(API_DOMAIN)
+                                     .addChild(CUSTOM_DOMAIN_BASE_PATH)
+                                     .addChild(SERIAL_PUBLICATION_PATH)
+                                     .getUri();
         this.channelRegistryPathElement = "/findjournalserie/";
     }
