+++ src/test/java/no/sikt/nva/pubchannels/handler/fetch/series/FetchSeriesByIdentifierAndYearHandlerTest.java
@@ -1,8 +1,9 @@
 package no.sikt.nva.pubchannels.handler.fetch.series;
 
+import static no.sikt.nva.pubchannels.TestConstants.API_DOMAIN;
+import static no.sikt.nva.pubchannels.TestConstants.CUSTOM_DOMAIN_BASE_PATH;
 import static no.sikt.nva.pubchannels.TestConstants.SERIES_PATH;
 import static no.sikt.nva.pubchannels.TestConstants.SERIES_TYPE;
-import java.net.URI;
 import no.sikt.nva.pubchannels.channelregistry.ChannelRegistryClient;
 import no.sikt.nva.pubchannels.channelregistrycache.db.service.CacheService;
 import no.sikt.nva.pubchannels.handler.PublicationChannelClient;
@@ -11,6 +12,7 @@ import no.sikt.nva.pubchannels.handler.fetch.FetchByIdentifierAndYearHandler;
 import no.sikt.nva.pubchannels.handler.model.SerialPublicationDto;
 import no.sikt.nva.pubchannels.utils.AppConfig;
 import nva.commons.core.Environment;
+import nva.commons.core.paths.UriWrapper;
 import org.junit.jupiter.api.BeforeEach;
 
 class FetchSeriesByIdentifierAndYearHandlerTest extends BaseFetchSerialPublicationByIdentifierAndYearHandlerTest {
@@ -39,7 +41,10 @@ class FetchSeriesByIdentifierAndYearHandlerTest extends BaseFetchSerialPublicati
                                                                           super.getAppConfigWithCacheEnabled(false));
         this.type = SERIES_TYPE;
         this.customChannelPath = SERIES_PATH;
-        this.selfBaseUri = URI.create("https://localhost/publication-channels/" + SERIES_PATH);
+        this.selfBaseUri = UriWrapper.fromHost(API_DOMAIN)
+                                     .addChild(CUSTOM_DOMAIN_BASE_PATH)
+                                     .addChild(SERIES_PATH)
+                                     .getUri();
         this.channelRegistryPathElement = "/findseries/";
     }
 }
